Check within half a mile of image location

+++ Images/Models/MessageDetails.cs
@@ -4,7 +4,7 @@
     {
         public string Image { get; set; }
         public string Message { get; set; }
-        public float? Latitude { get; set; }
-        public float? Longitude { get; set; }
+        public double? Latitude { get; set; }
+        public double? Longitude { get; set; }
     }
 }
